Do not link against liblua.so
Because lgi.so needs is a plugin, it will get all needed Lua symbols
from embedding module (typically interpreter), so it will be always
compatible with it.  Note that when win32 port will be created, this
will have to be a bit more complicated, because win32 loader has no
weak backward references.