Improve core.registerlock to handle re-opening of already opened library
This is important in the case that the lua_State with lgi loaded is
closed and then opeend again.  In this case, we have to avoid calling
register lock function, because it was already called before.

Also rewrite core.registerlock to take advantage of gi symbol resolver
instead of duplicating it here locally.

https://github.com/pavouk/lgi/issues/9