Fix various errors in the cairo bindings
I went through the list of cairo functions in override/cairo.lua and compared
each one with cairo's source code. This patch fixes the various inconsistencies
and errors that I spotted.

Also, I removed the 'since' attribute for the mesh pattern constructor, because
the all of the mesh patterns already have the right since attribute.

The binding for cairo_recording_surface_create() is still flawed.

There is a special behavior for its second argument being NULL. I don't know how
something like this can be expressed in lgi (it creates a new cairo_rectangle_t
when I call the function with a nil argument, doesn't it?).

Signed-off-by: Uli Schlachter <psychon@znc.in>