cairo: Add new properties to cairo contexts
This adds the group_target, matrix and font_matrix properties to cairo contexts.

This uses cairo_get_group_target() which is already handled correctly. This new
group_target property complements the existing target property, which is
read-only, too.

Similar, the setters and getters for matrix and font_matrix already exist and
can easily be turned into a property

Signed-off-by: Uli Schlachter <psychon@znc.in>