Allow overriding of the host build system
When cross compiling, the build system is different from the host
system, so directly using `uname` is not an option.

Use the helper variable HOST_OS so it can be overriden when calling
make and leave the current command (uname -s) as its default value.
Better would be to use a full system triplet and rely on config.guess
(like autoconf does) but this would significantly increase the
complexity of the build system.