Add a test for DataInputStream:read_line()
The end of the input is indicated in the C API by read_line() return a NULL
pointer (and this return value is correctly annotated as [nullable], but I don't
know what to do with that info). However, lgi will change this NULL into an
empty string which makes it indistinguishable from reading an empty line.

This commit adds a test for Gio that tests for the above behaviour and expects
the result to be "nil" when the end of the input is reached. Thus, this test
fails currently.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>