Fix stack handling in set_resident()
Commit 2a2b60644fed2cc made this function pop the nil value that lua_getfield()
might push on the stack earlier than before. However, the following code (run
when (lua_gettop(L)==2 fails) assumes that the nil value is still on the stack
so that lua_next() works correctly. This can also be seen in the comment before
the loop:

      /* Note: 'nil' is on the stack from lua_getfield() call above. */

The above commit broke this.

Fix this with a partial revert of the above commit to restore the previous stack
handling. The new code added in the above commit is changed to cope with the
extra value on the stack.

Signed-off-by: Uli Schlachter <psychon@znc.in>