Simplify Makefile: Don't query for default LUA_PATH
All Lua versions since at least Lua 5.1 replace ";;" in $LUA_PATH and $LUA_CPATH
with the default value. Use this instead of trying to query the default path so
that we can prepend something.

If $LUA_PATH is empty, we end up with ";;" in the new value, so the default path
is added. If $LUA_PATH is not empty, we add our own entry to it.

This also fixes a (theoretical) issue due to missing quoting on the path.

Signed-off-by: Uli Schlachter <psychon@znc.in>