Add a .travis.yml file
This allows automatically building and running tests on Travis for every commit
and pull request. Travis will build four different configurations which differ
in the Lua version that is being used: Lua 5.1, 5.2, 5.3 and LuaJIT are tested.

Lua 5.1 and Lua 5.2 are installed from the Ubuntu package repository.

Lua 5.3.3 and latest LuaJIT git are built manually. For Lua 5.3, additional
asserts are enabled to catch API mis-use (e.g. lua_pop() on an empty stack).

This .travis.yml is loosely based on the same file in awesomeWM's git
repository.

Signed-off-by: Uli Schlachter <psychon@znc.in>