Small refactoring in closure_callback
This function had a variable L for some Lua state and almost immediately
this value was overwritten with a value fetched from some Lua reference.
This was confusing me and lead me to write a wrong patch.

Refactor the code so that the value of this variable is not overwritten
right after being set. Hopefully this makes it easier to see where its
value comes from in the future.

Signed-off-by: Uli Schlachter <psychon@znc.in>