Fix possible crash when invoking a callback
The code for invoking a callback first tries to figure out which
coroutine to use for this. Here it is possible that a new coroutine is
created to do this call. Specifically, the code creates a new coroutine
(lua_newthread()) and then registers this one in the registry to prevent
it from being garbage collected prematurely.

However, the code was wrong. lua_newthread() produces a new coroutine,
pushes it onto the stack and returns its lua_State pointer. This new
lua_State pointer was used to overwrite the variable L here. This means
that the following line that tried to create a strong reference to the
coroutine instead called lua_rawseti() on an empty stack. In a debug
build of Lua (LUA_USE_APICHECK preprocessor symbol defined) this would
cause a failed assertion. In a normal build "something" happened where
"something" often ended in a segfault.

Fix this by using the correct Lua state in the argument to
lua_rawseti().

This commit also adds a regression test for this. The test counts to 100
in a complicated way involving a callback through lgi. In my
experiments, the test causes the crash after just two iterations.
However, earlier versions needed more iterations.  100 iterations only
take about 0.6 milliseconds here, so should still be fast enough not to
be about problem, but should still be good enough to detect the problem.

Fixes: https://github.com/pavouk/lgi/issues/155
Fixes: https://github.com/pavouk/lgi/issues/143
Signed-off-by: Uli Schlachter <psychon@znc.in>