.travis.yml: Use xvfb-run with argument "-a"
Some builds on Travis failed because the second call to xvfb-run failed
with "xvfb-run: error: Xvfb failed to start".

According to google, this happens when the display number is already in
use, for example from a previous call to xvfb-run. Since we recently
added a second call to xvfb-run to .travis.yml, this seems plausible to
me.

Fix this by calling xvfb-run with --auto-servernum (aka -a), which means
that xvfb-run automatically picks a free display.

Signed-off-by: Uli Schlachter <psychon@znc.in>