closure_callback: Use separate thread for marshalling
When a thread is yielded (status LUA_YIELD), Lua can get really upset
when one tries to execute Lua code on that thread. When LUA_USE_APICHECK
is enabled and e.g. things run under code coverage analysis (meaning
that debug.sethook() was used to set a hook that runs on various
occasions), this can actually result in a failed assertion which aborts
the process.

Fix this by making closure_callback create a new Lua thread and use that
one for all marshaling. Only the results are then moved back to the
yielded Lua thread's stack and the coroutine is resumed.

Sadly, I cannot provide a nice test case that shows the error before
this commit and the fix afterwards. I only got a complicated one.

Signed-off-by: Uli Schlachter <psychon@znc.in>