closure_callback: Use a separate thread for marshalling
When a thread is yielded (status LUA_YIELD), it is not allowed to run
any Lua code on it, not even through metamethods. See e.g. [1].

However, currently lgi breaks this rule when marshalling arguments and
return values for a yielded thread. Fix this by making
closure_callback() use a separate Lua thread for all marshalling. Only
the results are then moved to the actual target thread's stack.

I don't have any good test case for this that actually "breaks stuff",
so no new test is added.

[1]: http://lua-users.org/lists/lua-l/2017-11/msg00095.html

Signed-off-by: Uli Schlachter <psychon@znc.in>