cairo: Use the typelib before loading cairo directly
Doing core.module('cairo', 2) results in dlopen("libcairo.so.2"). This
causes a problem in NixOS since there is no single, global library
search path and instead absolute paths to libraries are inserted during
build. Thus, loading the library fails.

However, in NixOS the cairo-1.0.typelib is patched to contain a full
path to libcairo.so.2. Thus, when causing this library to be loaded
through gobject-introspection, everything works out.

To achieve this goal, this commit reorders some code in the cairo
override file so that first some enums are loaded through
gobject-introspection (causing libcairo.so to be loaded) and only
afterwards cairo is opened directly (which does not need a search path
since cairo was already loaded before).

Related to: https://github.com/NixOS/nixpkgs/pull/34584
Signed-off-by: Uli Schlachter <psychon@znc.in>