Add a new test case for multiple Lua states
Commit 44db274a894a669a added a special, internal thread that is used
for argument marshalling in case we have to resume a coroutine. In this
case, the target thread cannot be used for marshalling, because
marshalling can run Lua code and that is not allowed on a yielded
thread/coroutine.

However, the fix there added a global variable that is always used for
argument marshalling. This breaks if the program uses multiple, separate
Lua states, because then we would mix things up between these multiple
states.

This commit adds a test that exhibits this problem. The test currently
segfaults.

Signed-off-by: Uli Schlachter <psychon@znc.in>