meson: Run tests correctly
There was a hack where some extra arguments are passed to the test
runner by meson. This hack was necessary so that some dependencies are
generated before the test itself was started. However, these arguments
were interpreted as a list of subtests to run and so no tests were
actually run.

Fix this by explicitly specifying these targets as dependencies. This is
possible in newer versions of meson, hence the required meson version
gets bumped.

Fixes: https://github.com/pavouk/lgi/issues/214
Signed-off-by: Uli Schlachter <psychon@znc.in>