Fix compilation of the new test case on Travis
The new test case is compiled into a self-contained executable. Thus,
this now has to link against the right Lua libraries. This commit adds
that information to the linker invocation.

For users, this might become a bit of a problem. Right now, the Makefile
guesses -llua to link against Lua, but there are many computers out
there where this is not correct, but e.g. -llua5.2 is needed. Users will
either have to not run tests, or to provide the right value for $LUA_LIB
themselves. Sorry.

This also fixes the Lua 5.3 build so that liblua.so is linked correctly
against its dependencies

Signed-off-by: Uli Schlachter <psychon@znc.in>