Fix bytestring handling with embedded null bytes
This all started with me noticing that when transforming a bytestring to
Lua, the resulting string was truncated at the first null byte. I added
a new test variant.bytestring_embedded_null() (see diff) and started
trying to fix this test.

The fix is simple: Instead of using g_variant_get_bytestring() (which
only providers a pointer to the beginning of the data, but no length) to
transform the value to Lua, the code has to use the already existing
.data property. This code carefully gets the length of the bytestring
and creates a bytes object of appropriate length. Easy fix. This is
where the rabbit hole started.

The docs for g_variant_new_bytestring() say:

  The nul terminator character at the end of the string is stored in the
  array.

Thus, all existing code that creates a GVariant of type 'ay' from
strings actually include the terminating \0 byte. Since Lua strings have
explicit length, this behaviour is quite non-intuitive. To get a
GVariant bytestring without the terminating \0 byte, one possibility is
to use g_variant_new_from_data(). Thus, to fix some tests, this commit
changes the code to use that function for constructing GVariant objects
that really only contain what the given Lua string.

However, without a terminating \0 byte, g_variant_get_bytestring()
refuses to work and just returns a pointer to the empty string. This
broke the test variant.newv_basic(). To overcome that problem, the
GVariant used in the test was extended with a \0 byte.

The approach taken in this commit guarantees that converting from Lua
string -> GVariant bytestring -> Lua string produces the original Lua
string again. An alternative approach could be to just live with the
extra null byte that shows up. However, since that behaviour is quite
surprising, I went with the alternative approach instead: Getting rid of
this extra byte.

The approach I took however has a higher chance of causing breakage
elsewhere.

The original problem (.data truncating at the first embedded zero byte)
was first observed here (but the problem was not yet identified):
https://github.com/awesomeWM/awesome/pull/2722#issuecomment-475322079

Signed-off-by: Uli Schlachter <psychon@znc.in>