Do not mark callbacks as "internal"
Data bound to a callback must be flagged as "internal". The problem is
also the callback itself is marked as such because of this issue:

https://gitlab.gnome.org/GNOME/gobject-introspection/-/issues/430

Avoid the problem by checking the scope: if it is invalid, it is likely
real data (and not a callback), hence it can be marked.

Fixes #241, #262 and #285.