bitflags: Return 0 if param is nil
This allows the user to opt out of passing flags into functions when
there are none to specify. While libraries in statically typed languages
expect users to pass in an explicit _NONE flag, Lua's ability to pass a
variable number of parameters could easily be interpreted as explicitly
not giving any flags when none are passed.

Fixes #329